chore: rename all `.jsdom.test` files
