+++ packages/@ourworldindata/grapher/src/stackedCharts/MarimekkoChart.test.tsx
@@ -1,51 +1,21 @@
-import { expect, it, describe } from "vitest"
+/**
+ * @vitest-environment happy-dom
+ */
 
-import {
-    Bounds,
-    SortOrder,
-    SortBy,
-    ColumnTypeNames,
-} from "@ourworldindata/utils"
-import {
-    OwidTable,
-    SampleColumnSlugs,
-    SynthesizeGDPTable,
-} from "@ourworldindata/core-table"
+import { expect, it, describe, test } from "vitest"
+
+import { Bounds, ColumnTypeNames, omit } from "@ourworldindata/utils"
+import { OwidTable } from "@ourworldindata/core-table"
 import { DefaultColorScheme } from "../color/CustomSchemes"
+import { Grapher } from "../core/Grapher"
+import { GRAPHER_CHART_TYPES } from "@ourworldindata/types"
 import { MarimekkoChart } from "./MarimekkoChart"
-import {
-    MarimekkoChartManager,
-    BarShape,
-    Item,
-    PlacedItem,
-} from "./MarimekkoChartConstants"
-
-it("can create a chart", () => {
-    const table = SynthesizeGDPTable({
-        timeRange: [2000, 2001],
-        entityCount: 5,
-    })
-    const manager = {
-        table,
-        yColumnSlugs: [SampleColumnSlugs.GDP],
-        xColumnSlug: SampleColumnSlugs.Population,
-        showNoDataArea: false,
-    }
-
-    const chart = new MarimekkoChart({ manager })
-    //expect(chart.failMessage).toBeTruthy()
-
-    // selection.addToSelection(table.sampleEntityName(5))
-    expect(chart.failMessage).toEqual("")
-    expect(chart.series.length).toEqual(1)
-    expect(chart.series[0].points.length).toEqual(5)
-    expect(chart.xSeries!.points.length).toEqual(5)
-    expect(chart.placedItems.length).toEqual(5)
-    //expect(chart.placedItems)
-})
-
-it("can display a Marimekko chart correctly", () => {
+import { BarShape, PlacedItem } from "./MarimekkoChartConstants"
+it("can filter years correctly", () => {
     const csv = `year,entityName,population,percentBelow2USD
+2000,medium,4000,5
+2000,big,5000,10
+2000,small,800,2
 2001,medium,4000,4
 2001,big,5000,8
 2001,small,1000,3`
@@ -55,19 +25,22 @@ it("can display a Marimekko chart correctly", () => {
         { slug: "year", type: ColumnTypeNames.Year },
     ])
 
-    const manager: MarimekkoChartManager = {
+    // TODO: why is it ySlugs and xSlug here instead of yColumnSlugs and xColumnSlug? Unify when we have config migrations?
+    const manager = {
+        chartTypes: [GRAPHER_CHART_TYPES.Marimekko],
         table,
         selection: table.availableEntityNames,
-        yColumnSlugs: ["percentBelow2USD"],
-        xColumnSlug: "population",
+        ySlugs: "percentBelow2USD",
+        xSlug: "population",
         endTime: 2001,
-        showNoDataArea: false,
     }
+    const grapher = new Grapher(manager)
     const chart = new MarimekkoChart({
-        manager,
+        manager: grapher,
         bounds: new Bounds(0, 0, 1000, 1000),
     })
     const xAxisRange = chart["dualAxis"].horizontalAxis.rangeSize
+    //grapher.startHandleTimeBound = 2000
 
     expect(chart.failMessage).toEqual("")
     expect(chart.series.length).toEqual(1)
@@ -148,51 +121,41 @@ it("can display a Marimekko chart correctly", () => {
     ])
 })
 
-function roundXPosition(item: PlacedItem): PlacedItem {
-    return {
-        ...item,
-        xPosition: Math.round(item.xPosition),
-    }
-}
-
-it("can display two time series stacked correctly", () => {
-    const csv = `year,entityName,population,percentBelow2USD,percentBelow10USD
-2001,medium,4000,4,8.5
-2001,big,5000,8,20
-2001,small,1000,3,5`
+it("shows no data points at the end", () => {
+    const csv = `year,entityName,population,percentBelow2USD
+2000,medium,4000,5
+2000,big,5000,10
+2000,small,800,2
+2001,medium,4000,
+2001,big,5000,8
+2001,small,1000,3`
     const table = new OwidTable(csv, [
         { slug: "population", type: ColumnTypeNames.Numeric },
         { slug: "percentBelow2USD", type: ColumnTypeNames.Numeric },
-        { slug: "percentBelow10USD", type: ColumnTypeNames.Numeric },
         { slug: "year", type: ColumnTypeNames.Year },
     ])
 
-    const manager: MarimekkoChartManager = {
+    // TODO: why is it ySlugs and xSlug here instead of yColumnSlugs and xColumnSlug? Unify when we have config migrations?
+    const manager = {
+        chartTypes: [GRAPHER_CHART_TYPES.Marimekko],
         table,
         selection: table.availableEntityNames,
-        yColumnSlugs: ["percentBelow2USD", "percentBelow10USD"],
-        xColumnSlug: "population",
+        ySlugs: "percentBelow2USD",
+        xSlug: "population",
         endTime: 2001,
-        showNoDataArea: false,
     }
+    const grapher = new Grapher(manager)
     const chart = new MarimekkoChart({
-        manager,
-        bounds: new Bounds(0, 0, 1000, 1000),
+        manager: grapher,
+        bounds: new Bounds(0, 0, 1001, 1000),
     })
     const xAxisRange = chart["dualAxis"].horizontalAxis.rangeSize
 
     expect(chart.failMessage).toEqual("")
-    expect(chart.series.length).toEqual(2)
-    expect(chart.series[0].points.length).toEqual(3)
-    expect(chart.series[1].points.length).toEqual(3)
-    // Y series points should be one series in order of the data
-    const expectedYPointsFirstSeries = [
-        {
-            position: "medium",
-            value: 4,
-            valueOffset: 0,
-            time: 2001,
-        },
+    expect(chart.series.length).toEqual(1)
+    expect(chart.series[0].points.length).toEqual(2)
+    // Y series points should be one series in order of the data. Medium should be missing here
+    const expectedYPoints = [
         {
             position: "big",
             value: 8,
@@ -206,34 +169,13 @@ it("can display two time series stacked correctly", () => {
             time: 2001,
         },
     ]
-    const expectedYPointsSecondSeries = [
-        {
-            position: "medium",
-            value: 8.5,
-            valueOffset: 4,
-            time: 2001,
-        },
-        {
-            position: "big",
-            value: 20,
-            valueOffset: 8,
-            time: 2001,
-        },
-        {
-            position: "small",
-            value: 5,
-            valueOffset: 3,
-            time: 2001,
-        },
-    ]
     // X series points should be in order of the data
     const expectedXPoints = [
         { value: 4000, entity: "medium", time: 2001 },
         { value: 5000, entity: "big", time: 2001 },
         { value: 1000, entity: "small", time: 2001 },
     ]
-    expect(chart.series[0].points).toEqual(expectedYPointsFirstSeries)
-    expect(chart.series[1].points).toEqual(expectedYPointsSecondSeries)
+    expect(chart.series[0].points).toEqual(expectedYPoints)
     expect(chart.xSeries!.points).toEqual(expectedXPoints)
     // placedItems should be in default sort order
     expect(chart.placedItems.map(roundXPosition)).toEqual([
@@ -245,16 +187,117 @@ it("can display two time series stacked correctly", () => {
                     kind: BarShape.Bar,
                     color: DefaultColorScheme.colorSets[0][0],
                     seriesName: "percentBelow2USD",
-                    yPoint: expectedYPointsFirstSeries[1],
-                },
-                {
-                    kind: BarShape.Bar,
-                    color: DefaultColorScheme.colorSets[0][1],
-                    seriesName: "percentBelow10USD",
-                    yPoint: expectedYPointsSecondSeries[1],
+                    yPoint: expectedYPoints[0],
                 },
             ],
             xPoint: expectedXPoints[1],
+            xPosition: 0,
+        },
+        {
+            entityName: "small",
+            entityColor: undefined,
+            bars: [
+                {
+                    kind: BarShape.Bar,
+                    color: DefaultColorScheme.colorSets[0][0],
+                    seriesName: "percentBelow2USD",
+                    yPoint: expectedYPoints[1],
+                },
+            ],
+            xPoint: expectedXPoints[2],
+            xPosition: Math.round(xAxisRange * 0.5),
+        },
+        {
+            entityName: "medium",
+            entityColor: undefined,
+            bars: [],
+            xPoint: expectedXPoints[0],
+            xPosition: Math.round(xAxisRange * 0.6),
+        },
+    ])
+})
+
+test("interpolation works as expected", () => {
+    const csv = `year,entityName,population,percentBelow2USD
+2000,medium,4000,5
+2000,big,5000,10
+2000,small,800,2
+2001,medium,4000,4
+2001,big,,8
+2001,small,1000,`
+    const table = new OwidTable(csv, [
+        { slug: "population", type: ColumnTypeNames.Numeric, tolerance: 1 },
+        {
+            slug: "percentBelow2USD",
+            type: ColumnTypeNames.Numeric,
+            tolerance: 1,
+        },
+        { slug: "year", type: ColumnTypeNames.Year },
+    ])
+
+    // TODO: why is it ySlugs and xSlug here instead of yColumnSlugs and xColumnSlug? Unify when we have config migrations?
+    const manager = {
+        chartTypes: [GRAPHER_CHART_TYPES.Marimekko],
+        table,
+        selection: table.availableEntityNames,
+        ySlugs: "percentBelow2USD",
+        xSlug: "population",
+        endTime: 2001,
+    }
+    const grapher = new Grapher(manager)
+    const chart = new MarimekkoChart({
+        manager: grapher,
+        bounds: new Bounds(0, 0, 1000, 1000),
+    })
+    const xAxisRange = chart["dualAxis"].horizontalAxis.rangeSize
+
+    expect(chart.failMessage).toEqual("")
+    expect(chart.series.length).toEqual(1)
+    expect(chart.series[0].points.length).toEqual(3)
+    // Y series points should be one series in order of the data
+    const expectedYPoints = [
+        {
+            position: "big",
+            value: 8,
+            valueOffset: 0,
+            time: 2001,
+        },
+        {
+            position: "medium",
+            value: 4,
+            valueOffset: 0,
+            time: 2001,
+        },
+
+        {
+            position: "small",
+            value: 2,
+            valueOffset: 0,
+            time: 2000,
+        },
+    ]
+    // X series points should be in order of the data
+    const expectedXPoints = [
+        { value: 5000, entity: "big", time: 2000 },
+        { value: 4000, entity: "medium", time: 2001 },
+        { value: 1000, entity: "small", time: 2001 },
+    ]
+    expect(chart.series[0].points).toEqual(expectedYPoints)
+    expect(chart.xSeries!.points).toEqual(expectedXPoints)
+    // placedItems should be in default sort order
+    expect(chart.placedItems.map(roundXPosition)).toEqual([
+        {
+            entityName: "big",
+            entityColor: undefined,
+            bars: [
+                {
+                    kind: BarShape.Bar,
+                    color: DefaultColorScheme.colorSets[0][0],
+                    seriesName: "percentBelow2USD",
+                    yPoint: expectedYPoints[0],
+                },
+            ],
+            xPoint: expectedXPoints[0],
             xPosition: 0,
         },
         {
@@ -265,16 +308,10 @@ it("can display two time series stacked correctly", () => {
                     kind: BarShape.Bar,
                     color: DefaultColorScheme.colorSets[0][0],
                     seriesName: "percentBelow2USD",
-                    yPoint: expectedYPointsFirstSeries[0],
-                },
-                {
-                    kind: BarShape.Bar,
-                    color: DefaultColorScheme.colorSets[0][1],
-                    seriesName: "percentBelow10USD",
-                    yPoint: expectedYPointsSecondSeries[0],
+                    yPoint: expectedYPoints[1],
                 },
             ],
-            xPoint: expectedXPoints[0],
+            xPoint: expectedXPoints[1],
             xPosition: Math.round(xAxisRange * 0.5),
         },
         {
@@ -285,13 +322,7 @@ it("can display two time series stacked correctly", () => {
                     kind: BarShape.Bar,
                     color: DefaultColorScheme.colorSets[0][0],
                     seriesName: "percentBelow2USD",
-                    yPoint: expectedYPointsFirstSeries[2],
-                },
-                {
-                    kind: BarShape.Bar,
-                    color: DefaultColorScheme.colorSets[0][1],
-                    seriesName: "percentBelow10USD",
-                    yPoint: expectedYPointsSecondSeries[2],
+                    yPoint: expectedYPoints[2],
                 },
             ],
             xPoint: expectedXPoints[2],
@@ -300,158 +331,157 @@ it("can display two time series stacked correctly", () => {
     ])
 })
 
-it("can do sorting", () => {
-    const csv = `year,entityName,population,percentBelow2USD
-2001,AA,4000,4
-2001,BB,5000,8
-2001,CC,1000,3`
+it("can deal with y columns with missing values", () => {
+    const csv = `year,entityName,population,percentBelow2USD,percentBelow10USD
+2000,medium,4000,,10
+2000,big,5000,10,20
+2000,small,800,2,4
+2001,medium,4000,4,8
+2001,big,5000,8,
+2001,small,1000,3,6`
     const table = new OwidTable(csv, [
         { slug: "population", type: ColumnTypeNames.Numeric },
         { slug: "percentBelow2USD", type: ColumnTypeNames.Numeric },
+        { slug: "percentBelow10USD", type: ColumnTypeNames.Numeric },
         { slug: "year", type: ColumnTypeNames.Year },
     ])
 
-    const manager: MarimekkoChartManager = {
+    // TODO: why is it ySlugs and xSlug here instead of yColumnSlugs and xColumnSlug? Unify when we have config migrations?
+    const manager = {
+        chartTypes: [GRAPHER_CHART_TYPES.Marimekko],
         table,
         selection: table.availableEntityNames,
-        yColumnSlugs: ["percentBelow2USD"],
-        xColumnSlug: "population",
+        ySlugs: "percentBelow2USD percentBelow10USD",
+        xSlug: "population",
         endTime: 2001,
-        showNoDataArea: false,
     }
-    let chart = new MarimekkoChart({
-        manager: {
-            ...manager,
-            sortConfig: {
-                sortBy: SortBy.total,
-                sortOrder: SortOrder.asc,
-            },
-        },
-
+    const grapher = new Grapher(manager)
+    const chart = new MarimekkoChart({
+        manager: grapher,
         bounds: new Bounds(0, 0, 1000, 1000),
     })
+    const xAxisRange = chart["dualAxis"].horizontalAxis.rangeSize
+    //grapher.startHandleTimeBound = 2000
 
     expect(chart.failMessage).toEqual("")
-    expect(chart.series.length).toEqual(1)
+    expect(chart.series.length).toEqual(2)
     expect(chart.series[0].points.length).toEqual(3)
+    expect(chart.series[1].points.length).toEqual(2)
     // Y series points should be one series in order of the data
-    const expectedYPoints = [
+    const expectedYPoints1 = [
         {
-            position: "AA",
+            position: "medium",
             value: 4,
             valueOffset: 0,
             time: 2001,
         },
         {
-            position: "BB",
+            position: "big",
             value: 8,
             valueOffset: 0,
             time: 2001,
         },
         {
-            position: "CC",
+            position: "small",
             value: 3,
             valueOffset: 0,
             time: 2001,
         },
     ]
+    const expectedYPoints2 = [
+        {
+            position: "medium",
+            value: 8,
+            valueOffset: 4,
+            time: 2001,
+        },
+
+        {
+            position: "small",
+            value: 6,
+            valueOffset: 3,
+            time: 2001,
+        },
+    ]
     // X series points should be in order of the data
     const expectedXPoints = [
-        { value: 4000, entity: "AA", time: 2001 },
-        { value: 5000, entity: "BB", time: 2001 },
-        { value: 1000, entity: "CC", time: 2001 },
+        { value: 4000, entity: "medium", time: 2001 },
+        { value: 5000, entity: "big", time: 2001 },
+        { value: 1000, entity: "small", time: 2001 },
     ]
-    expect(chart.series[0].points).toEqual(expectedYPoints)
+    expect(chart.series[0].points).toEqual(expectedYPoints1)
+    expect(chart.series[1].points).toEqual(expectedYPoints2)
     expect(chart.xSeries!.points).toEqual(expectedXPoints)
+
+    const placedItemsWithoutXPosition = chart.placedItems.map((placedItem) =>
+        omit(placedItem, "xPosition")
+    )
+    const xPositions = chart.placedItems.map(
+        (placedItem) => placedItem.xPosition
+    )
     // placedItems should be in default sort order
-    const items = new Map<string, Item>([
-        [
-            "big",
-            {
-                entityName: "BB",
-                entityColor: undefined,
-                bars: [
-                    {
-                        kind: BarShape.Bar,
-                        color: DefaultColorScheme.colorSets[0][0],
-                        seriesName: "percentBelow2USD",
-                        yPoint: expectedYPoints[1],
-                    },
-                ],
-                xPoint: expectedXPoints[1],
-            },
-        ],
-        [
-            "medium",
-            {
-                entityName: "AA",
-                entityColor: undefined,
-                bars: [
-                    {
-                        kind: BarShape.Bar,
-                        color: DefaultColorScheme.colorSets[0][0],
-                        seriesName: "percentBelow2USD",
-                        yPoint: expectedYPoints[0],
-                    },
-                ],
-                xPoint: expectedXPoints[0],
-            },
-        ],
-        [
-            "small",
-            {
-                entityName: "CC",
-                entityColor: undefined,
-                bars: [
-                    {
-                        kind: BarShape.Bar,
-                        color: DefaultColorScheme.colorSets[0][0],
-                        seriesName: "percentBelow2USD",
-                        yPoint: expectedYPoints[2],
-                    },
-                ],
-                xPoint: expectedXPoints[2],
-            },
-        ],
-    ])
-    expect(chart["sortedItems"]).toEqual([
-        items.get("small"),
-        items.get("medium"),
-        items.get("big"),
-    ])
-
-    chart = new MarimekkoChart({
-        manager: {
-            ...manager,
-            sortConfig: {
-                sortBy: SortBy.column,
-                sortColumnSlug: "percentBelow2USD",
-                sortOrder: SortOrder.asc,
-            },
+    expect(placedItemsWithoutXPosition).toEqual([
+        {
+            entityName: "medium",
+            entityColor: undefined,
+            bars: [
+                {
+                    kind: BarShape.Bar,
+                    color: DefaultColorScheme.colorSets[0][0],
+                    seriesName: "percentBelow2USD",
+                    yPoint: expectedYPoints1[0],
+                },
+                {
+                    kind: BarShape.Bar,
+                    color: DefaultColorScheme.colorSets[0][1],
+                    seriesName: "percentBelow10USD",
+                    yPoint: expectedYPoints2[0],
+                },
+            ],
+            xPoint: expectedXPoints[0],
         },
-
-        bounds: new Bounds(0, 0, 1000, 1000),
-    })
-    expect(chart["sortedItems"]).toEqual([
-        items.get("small"),
-        items.get("medium"),
-        items.get("big"),
-    ])
-
-    chart = new MarimekkoChart({
-        manager: {
-            ...manager,
-            sortConfig: {
-                sortBy: SortBy.entityName,
-                sortOrder: SortOrder.asc,
-            },
+        {
+            entityName: "small",
+            entityColor: undefined,
+            bars: [
+                {
+                    kind: BarShape.Bar,
+                    color: DefaultColorScheme.colorSets[0][0],
+                    seriesName: "percentBelow2USD",
+                    yPoint: expectedYPoints1[2],
+                },
+                {
+                    kind: BarShape.Bar,
+                    color: DefaultColorScheme.colorSets[0][1],
+                    seriesName: "percentBelow10USD",
+                    yPoint: expectedYPoints2[1],
+                },
+            ],
+            xPoint: expectedXPoints[2],
         },
-
-        bounds: new Bounds(0, 0, 1000, 1000),
-    })
-    expect(chart["sortedItems"]).toEqual([
-        items.get("medium"),
-        items.get("big"),
-        items.get("small"),
+        {
+            entityName: "big",
+            entityColor: undefined,
+            bars: [
+                {
+                    kind: BarShape.Bar,
+                    color: DefaultColorScheme.colorSets[0][0],
+                    seriesName: "percentBelow2USD",
+                    yPoint: expectedYPoints1[1],
+                },
+            ],
+            xPoint: expectedXPoints[1],
+        },
     ])
+
+    expect(xPositions[0]).toEqual(0)
+    expect(xPositions[1]).toBeCloseTo(xAxisRange * 0.4, 0)
+    expect(xPositions[2]).toBeCloseTo(xAxisRange * 0.5, 0)
 })
+
+function roundXPosition(item: PlacedItem): PlacedItem {
+    return {
+        ...item,
+        xPosition: Math.round(item.xPosition),
+    }
+}
